+++ pytorch_lightning/metrics/converters.py
@@ -4,7 +4,6 @@ conversion to/from :class:`numpy.ndarray` and :class:`torch.Tensor` as well as u
 sync tensors between different processes in a DDP scenario, when needed.
 """
 
-import sys
 import numbers
 from typing import Union, Any, Callable, Optional
 
@@ -18,12 +17,13 @@ from pytorch_lightning.utilities.apply_func import apply_to_collection
 def _apply_to_inputs(func_to_apply: Callable, *dec_args, **dec_kwargs) -> Callable:
     """
     Decorator function to apply a function to all inputs of a function.
+
     Args:
         func_to_apply: the function to apply to the inputs
         *dec_args: positional arguments for the function to be applied
         **dec_kwargs: keyword arguments for the function to be applied
 
-    Returns:
+    Return:
         the decorated function
     """
 
@@ -42,12 +42,13 @@ def _apply_to_inputs(func_to_apply: Callable, *dec_args, **dec_kwargs) -> Callab
 def _apply_to_outputs(func_to_apply: Callable, *dec_args, **dec_kwargs) -> Callable:
     """
     Decorator function to apply a function to all outputs of a function.
+
     Args:
         func_to_apply: the function to apply to the outputs
         *dec_args: positional arguments for the function to be applied
         **dec_kwargs: keyword arguments for the function to be applied
 
-    Returns:
+    Return:
         the decorated function
     """
 
@@ -69,9 +70,8 @@ def _convert_to_tensor(data: Any) -> Any:
     Args:
         data: the data to convert to tensor
 
-    Returns:
+    Return:
         the converted data
-
     """
     if isinstance(data, numbers.Number):
         return torch.tensor([data])
@@ -86,12 +86,12 @@ def _convert_to_tensor(data: Any) -> Any:
 
 def _convert_to_numpy(data: Union[torch.Tensor, np.ndarray, numbers.Number]) -> np.ndarray:
     """Convert all tensors and numpy arrays to numpy arrays.
+
     Args:
         data: the tensor or array to convert to numpy
 
-    Returns:
+    Return:
         the resulting numpy array
-
     """
     if isinstance(data, torch.Tensor):
         return data.cpu().detach().numpy()
@@ -103,6 +103,33 @@ def _convert_to_numpy(data: Union[torch.Tensor, np.ndarray, numbers.Number]) -> 
     raise TypeError("The given type ('%s') cannot be converted to a numpy array!" % type(data).__name__)
 
 
+def _numpy_metric_input_conversion(func_to_decorate: Callable) -> Callable:
+    """
+    Decorator converting all inputs of a function to numpy
+
+    Args:
+        func_to_decorate: the function whose inputs shall be converted
+
+    Return:
+        Callable: the decorated function
+    """
+    return _apply_to_inputs(
+        apply_to_collection, (torch.Tensor, np.ndarray, numbers.Number), _convert_to_numpy)(func_to_decorate)
+
+
+def _tensor_metric_output_conversion(func_to_decorate: Callable) -> Callable:
+    """
+    Decorator converting all outputs of a function to tensors
+
+    Args:
+        func_to_decorate: the function whose outputs shall be converted
+
+    Return:
+        Callable: the decorated function
+    """
+    return _apply_to_outputs(_convert_to_tensor)(func_to_decorate)
+
+
 def _numpy_metric_conversion(func_to_decorate: Callable) -> Callable:
     """
     Decorator handling the argument conversion for metrics working on numpy.
@@ -112,17 +139,43 @@ def _numpy_metric_conversion(func_to_decorate: Callable) -> Callable:
     Args:
         func_to_decorate: the function whose inputs and outputs shall be converted
 
-    Returns:
+    Return:
         the decorated function
-
     """
     # applies collection conversion from tensor to numpy to all inputs
     # we need to include numpy arrays here, since otherwise they will also be treated as sequences
-    func_convert_inputs = _apply_to_inputs(
-        apply_to_collection, (torch.Tensor, np.ndarray, numbers.Number), _convert_to_numpy)(func_to_decorate)
+    func_convert_inputs = _numpy_metric_input_conversion(func_to_decorate)
     # converts all inputs back to tensors (device doesn't matter here, since this is handled by BaseMetric)
-    func_convert_in_out = _apply_to_outputs(_convert_to_tensor)(func_convert_inputs)
+    func_convert_in_out = _tensor_metric_output_conversion(func_convert_inputs)
     return func_convert_in_out
+
+
+def _tensor_metric_input_conversion(func_to_decorate: Callable) -> Callable:
+    """
+    Decorator converting all inputs of a function to tensors
+
+    Args:
+        func_to_decorate: the function whose inputs shall be converted
+
+    Return:
+        Callable: the decorated function
+    """
+    return _apply_to_inputs(
+        apply_to_collection, (torch.Tensor, np.ndarray, numbers.Number), _convert_to_tensor)(func_to_decorate)
+
+
+def _tensor_collection_metric_output_conversion(func_to_decorate: Callable) -> Callable:
+    """
+    Decorator converting all numpy arrays and numbers occuring in the outputs of a function to tensors
+
+    Args:
+        func_to_decorate: the function whose outputs shall be converted
+
+    Return:
+        Callable: the decorated function
+    """
+    return _apply_to_outputs(apply_to_collection, (torch.Tensor, np.ndarray, numbers.Number),
+                             _convert_to_tensor)(func_to_decorate)
 
 
 def _tensor_metric_conversion(func_to_decorate: Callable) -> Callable:
@@ -133,16 +186,33 @@ def _tensor_metric_conversion(func_to_decorate: Callable) -> Callable:
     Args:
         func_to_decorate: the function whose inputs and outputs shall be converted
 
-    Returns:
+    Return:
         the decorated function
-
     """
     # converts all inputs to tensor if possible
     # we need to include tensors here, since otherwise they will also be treated as sequences
-    func_convert_inputs = _apply_to_inputs(
-        apply_to_collection, (torch.Tensor, np.ndarray, numbers.Number), _convert_to_tensor)(func_to_decorate)
+    func_convert_inputs = _tensor_metric_input_conversion(func_to_decorate)
     # convert all outputs to tensor if possible
-    return _apply_to_outputs(_convert_to_tensor)(func_convert_inputs)
+    return _tensor_metric_output_conversion(func_convert_inputs)
+
+
+def _tensor_collection_metric_conversion(func_to_decorate: Callable) -> Callable:
+    """
+    Decorator Handling the argument conversion for metrics working on tensors.
+    All inputs of the decorated function and all numpy arrays and numbers in
+    it's outputs will be converted to tensors
+
+    Args:
+        func_to_decorate: the function whose inputs and outputs shall be converted
+
+    Return:
+        the decorated function
+    """
+    # converts all inputs to tensor if possible
+    # we need to include tensors here, since otherwise they will also be treated as sequences
+    func_convert_inputs = _tensor_metric_input_conversion(func_to_decorate)
+    # convert all outputs to tensor if possible
+    return _tensor_collection_metric_output_conversion(func_convert_inputs)
 
 
 def _sync_ddp_if_available(result: Union[torch.Tensor],
@@ -157,9 +227,8 @@ def _sync_ddp_if_available(result: Union[torch.Tensor],
         group: the process group to gather results from. Defaults to all processes (world)
         reduce_op: the reduction operation. Defaults to sum.
 
-    Returns:
+    Return:
         reduced value
-
     """
 
     if torch.distributed.is_available() and torch.distributed.is_initialized():
@@ -177,11 +246,32 @@ def _sync_ddp_if_available(result: Union[torch.Tensor],
     return result
 
 
+def sync_ddp(group: Optional[Any] = None,
+             reduce_op: Optional[torch.distributed.ReduceOp] = None) -> Callable:
+    """
+    This decorator syncs a functions outputs across different processes for DDP.
+
+    Args:
+        group: the process group to gather results from. Defaults to all processes (world)
+        reduce_op: the reduction operation. Defaults to sum
+
+    Return:
+        the decorated function
+
+    """
+
+    def decorator_fn(func_to_decorate):
+        return _apply_to_outputs(apply_to_collection, torch.Tensor,
+                                 _sync_ddp_if_available, group=group,
+                                 reduce_op=reduce_op)(func_to_decorate)
+
+    return decorator_fn
+
+
 def numpy_metric(group: Optional[Any] = None,
                  reduce_op: Optional[torch.distributed.ReduceOp] = None) -> Callable:
     """
     This decorator shall be used on all function metrics working on numpy arrays.
-
     It handles the argument conversion and DDP reduction for metrics working on numpy.
     All inputs of the decorated function will be converted to numpy and all
     outputs will be converted to tensors.
@@ -191,15 +281,12 @@ def numpy_metric(group: Optional[Any] = None,
         group: the process group to gather results from. Defaults to all processes (world)
         reduce_op: the reduction operation. Defaults to sum
 
-    Returns:
+    Return:
         the decorated function
-
     """
 
     def decorator_fn(func_to_decorate):
-        return _apply_to_outputs(apply_to_collection, torch.Tensor, _sync_ddp_if_available,
-                                 group=group,
-                                 reduce_op=reduce_op)(_numpy_metric_conversion(func_to_decorate))
+        return sync_ddp(group=group, reduce_op=reduce_op)(_numpy_metric_conversion(func_to_decorate))
 
     return decorator_fn
 
@@ -208,7 +295,6 @@ def tensor_metric(group: Optional[Any] = None,
                   reduce_op: Optional[torch.distributed.ReduceOp] = None) -> Callable:
     """
     This decorator shall be used on all function metrics working on tensors.
-
     It handles the argument conversion and DDP reduction for metrics working on tensors.
     All inputs and outputs of the decorated function will be converted to tensors.
     In DDP Training all output tensors will be reduced according to the given rules.
@@ -217,14 +303,34 @@ def tensor_metric(group: Optional[Any] = None,
        group: the process group to gather results from. Defaults to all processes (world)
        reduce_op: the reduction operation. Defaults to sum
 
-    Returns:
+    Return:
        the decorated function
-
     """
 
     def decorator_fn(func_to_decorate):
-        return _apply_to_outputs(apply_to_collection, torch.Tensor, _sync_ddp_if_available,
-                                 group=group,
-                                 reduce_op=reduce_op)(_tensor_metric_conversion(func_to_decorate))
+        return sync_ddp(group=group, reduce_op=reduce_op)(_tensor_metric_conversion(func_to_decorate))
+
+    return decorator_fn
+
+
+def tensor_collection_metric(group: Optional[Any] = None,
+                             reduce_op: Optional[torch.distributed.ReduceOp] = None) -> Callable:
+    """
+    This decorator shall be used on all function metrics working on tensors and returning collections
+    that cannot be converted to tensors.
+    It handles the argument conversion and DDP reduction for metrics working on tensors.
+    All inputs and outputs of the decorated function will be converted to tensors.
+    In DDP Training all output tensors will be reduced according to the given rules.
+
+    Args:
+       group: the process group to gather results from. Defaults to all processes (world)
+       reduce_op: the reduction operation. Defaults to sum
+
+    Return:
+       the decorated function
+    """
+
+    def decorator_fn(func_to_decorate):
+        return sync_ddp(group=group, reduce_op=reduce_op)(_tensor_collection_metric_conversion(func_to_decorate))
 
     return decorator_fn