+++ pytorch_lightning/trainer/__init__.py
@@ -585,6 +585,19 @@ Example::
     --env=XLA_USE_BF16=1
     -- python your_trainer_file.py
 
+prepare_data_per_node
+^^^^^^^^^^^^^^^^^^^^^
+If True will call `prepare_data()` on LOCAL_RANK=0 for every node.
+If False will only call from NODE_RANK=0, LOCAL_RANK=0
+
+Example::
+
+    # default
+    Trainer(prepare_data_per_node=True)
+
+    # use only NODE_RANK=0, LOCAL_RANK=0
+    Trainer(prepare_data_per_node=False)
+
 tpu_cores
 ^^^^^^^^^
 - How many TPU cores to train on (1 or 8).
@@ -793,7 +806,7 @@ How often to add logging rows (does not write to disk)
 Example::
 
     # default used by the Trainer
-    trainer = Trainer(row_log_interval=10)
+    trainer = Trainer(row_log_interval=50)
 
 use_amp: