+++ pytorch_lightning/trainer/training_io.py
@@ -87,12 +87,12 @@ import os
 import re
 import signal
 from abc import ABC
-from argparse import Namespace
 from subprocess import call
 
 import torch
 import torch.distributed as torch_distrib
 
+import pytorch_lightning
 from pytorch_lightning import _logger as log
 from pytorch_lightning.core.lightning import LightningModule
 from pytorch_lightning.loggers import LightningLoggerBase
@@ -100,7 +100,7 @@ from pytorch_lightning.overrides.data_parallel import (
     LightningDistributedDataParallel,
     LightningDataParallel,
 )
-from pytorch_lightning.utilities import rank_zero_warn, parsing
+from pytorch_lightning.utilities import rank_zero_warn
 from pytorch_lightning.utilities.io import load as pl_load
 
 try:
@@ -119,6 +119,11 @@ except ImportError:
 else:
     HOROVOD_AVAILABLE = True
 
+try:
+    from omegaconf import Container
+except ImportError:
+    Container = None
+
 
 class TrainerIOMixin(ABC):
 
@@ -132,7 +137,7 @@ class TrainerIOMixin(ABC):
     use_ddp2: bool
     use_horovod: bool
     checkpoint_callback: ...
-    proc_rank: int
+    global_rank: int
     weights_save_path: str
     logger: LightningLoggerBase
     early_stop_callback: ...
@@ -213,7 +218,7 @@ class TrainerIOMixin(ABC):
             signal.signal(signal.SIGTERM, self.term_handler)
 
     def sig_handler(self, signum, frame):  # pragma: no-cover
-        if self.proc_rank == 0:
+        if self.is_global_zero:
             # save weights
             log.info('handling SIGUSR1')
             self.hpc_save(self.weights_save_path, self.logger)
@@ -262,13 +267,13 @@ class TrainerIOMixin(ABC):
     def save_checkpoint(self, filepath, weights_only: bool = False):
         checkpoint = self.dump_checkpoint(weights_only)
 
-        if self.proc_rank == 0:
+        if self.is_global_zero:
             # do the actual save
             try:
                 self._atomic_save(checkpoint, filepath)
             except AttributeError as err:
-                if LightningModule.CHECKPOINT_KEY_HYPER_PARAMS in checkpoint:
-                    del checkpoint[LightningModule.CHECKPOINT_KEY_HYPER_PARAMS]
+                if LightningModule.CHECKPOINT_HYPER_PARAMS_KEY in checkpoint:
+                    del checkpoint[LightningModule.CHECKPOINT_HYPER_PARAMS_KEY]
                 rank_zero_warn('Warning, `module_arguments` dropped from checkpoint.'
                                f' An attribute is not picklable {err}')
                 self._atomic_save(checkpoint, filepath)
@@ -320,6 +325,7 @@ class TrainerIOMixin(ABC):
         checkpoint = {
             'epoch': self.current_epoch + 1,
             'global_step': self.global_step + 1,
+            'pytorch-ligthning_version': pytorch_lightning.__version__,
         }
 
         if not weights_only:
@@ -356,10 +362,12 @@ class TrainerIOMixin(ABC):
 
         if model.hparams:
             if hasattr(model, '_hparams_name'):
-                checkpoint[LightningModule.CHECKPOINT_NAME_HYPER_PARAMS] = model._hparams_name
+                checkpoint[LightningModule.CHECKPOINT_HYPER_PARAMS_NAME] = model._hparams_name
             # add arguments to the checkpoint
-            # todo: add some recursion in case of OmegaConf
-            checkpoint[LightningModule.CHECKPOINT_KEY_HYPER_PARAMS] = dict(model.hparams)
+            checkpoint[LightningModule.CHECKPOINT_HYPER_PARAMS_KEY] = model.hparams
+            if Container is not None:
+                if isinstance(model.hparams, Container):
+                    checkpoint[LightningModule.CHECKPOINT_HYPER_PARAMS_TYPE] = type(model.hparams)
 
         # give the model a chance to add a few things
         model.on_save_checkpoint(checkpoint)
@@ -473,8 +481,8 @@ class TrainerIOMixin(ABC):
         try:
             self._atomic_save(checkpoint, filepath)
         except AttributeError as err:
-            if LightningModule.CHECKPOINT_KEY_HYPER_PARAMS in checkpoint:
-                del checkpoint[LightningModule.CHECKPOINT_KEY_HYPER_PARAMS]
+            if LightningModule.CHECKPOINT_HYPER_PARAMS_KEY in checkpoint:
+                del checkpoint[LightningModule.CHECKPOINT_HYPER_PARAMS_KEY]
             rank_zero_warn('warning, `module_arguments` dropped from checkpoint.'
                            f' An attribute is not picklable {err}')
             self._atomic_save(checkpoint, filepath)